Fix Homebrew Cask package installations
Apparently, while `brew` can install cask packages by just specifying a
package fully with the tap name and the package name (i.e. something like
`caskroom/cask/my-package`), the Ansible Homebrew module cannot. (Actually,
it appears that it can, but won't detect that the package is installed, and
will fail trying to reinstall in that case.) So we have to use the Ansible
`homebrew_cask` module.