Configure Control Strip of Touch Bar
I frequently hit the right-most button on the Touch Bar accidentally when
trying to hit the Delete (Backspace) key. By default, this was pulling up Siri,
even though I have Siri disabled. I'm changing this to Spotlight instead,
which seems to be the most innocuous option.

The Control Strip normally contains 4 controls on the right side of the Touch
Bar. It can be expanded to the full width of the Touch Bar by hitting the left
arrow to the left of the 4 controls. It can also be replaced with regular
function keys on a per-app basis.

To manually change these, in System Preferences, go to the Keyboard section,
click on the Keyboards tab, then click the Customize Control Strip button.